Initial post to GitHub
I've created a basic program that allows you to integrate Purple Air monitors with TP-Link devices managed through Kasa.  When pushed to a server in the cloud, this program allows for devices, like air purifiers, lights, etc. to be turned off or on based on the Air Quality of a specific Purple Air Sensor.